requirements.txt
openai
langchain
huggingface_hub
torch
git+https://github.com/huggingface/transformers
bitsandbytes
accelerate
datasets
loralib
sentencepiece
llama-cpp-python
pyyaml
duckduckgo-search
beautifulsoup4
tiktoken
PyPDF2
faiss-cpu
sentence_transformers